requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
playwright>=1.40.0
mistralai>=1.0.0
python-dotenv>=1.0.0
elevenlabs>=2.37.0
pyaudio>=0.2.14
